requirements.txt
flask==3.0.0
line-bot-sdk==3.5.0
beautifulsoup4==4.12.2
requests==2.31.0
gunicorn==21.2.0
